# Use types as context keys