--- a/context_test.go
+++ b/context_test.go
@@ -2,7 +2,11 @@ package rx
 
 import (
 	"math/rand"
+	"net/netip"
+	"reflect"
 	"testing"
+
+	"github.com/google/go-cmp/cmp"
 )
 
 func BenchmarkContextValues(b *testing.B) {
@@ -28,29 +32,48 @@ func BenchmarkContextValues(b *testing.B) {
 	// simulation note: pattern acces to values is usually skewed towards recent values
 	// to reproduce this, we use the order in switch, making sure that, statistically, key5 is modified twice each time key4 is modified, and so on…
 
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
+		ctx := Context{vx: &vctx{kv: make(map[reflect.Type]any)}}
 
-		orderkeys := rand.Perm(6)
-		ctx := Context{vx: &vctx{kv: make(map[ContextKey]any)}}
-
-		for i := 0; i < 30_000; i++ {
+		for range 30_000 {
 			rnd := rand.Uint32()
 			va := values[rnd%16]
 
 			switch {
 			case rnd&(1<<key5) > 0:
-				ctx = WithValue(ctx, ContextKey(key5+uint32(orderkeys[i%6])), va)
+				ctx = WithValue(ctx, va)
 			case rnd&(1<<key4) > 0:
-				ctx = WithValue(ctx, ContextKey(key4+uint32(orderkeys[i%6])), va)
+				ctx = WithValue(ctx, va)
 			case rnd&(1<<key3) > 0:
-				ctx = WithValue(ctx, ContextKey(key3+uint32(orderkeys[i%6])), va)
+				ctx = WithValue(ctx, va)
 			case rnd&(1<<key2) > 0:
-				ctx = WithValue(ctx, ContextKey(key2+uint32(orderkeys[i%6])), va)
+				ctx = WithValue(ctx, va)
 			case rnd&key1 > 0:
-				ctx = WithValue(ctx, ContextKey(key1+uint32(orderkeys[i%6])), va)
+				ctx = WithValue(ctx, va)
 			}
 
 			b.ReportMetric(float64(len(ctx.vx.kv)), "valuelength")
 		}
+	}
+}
+
+func TestMutator(t *testing.T) {
+	var ctx Context
+
+	type User struct{ name string }
+	type Endpoint int
+	type Dst netip.Addr
+
+	ctx = WithValues(ctx, User{name: "Doe"}, Endpoint(10))
+
+	Mutate(
+		func(u *User) { u.name = "Bond" },
+		func(e *Endpoint) { *e = 10 },
+		func(d *Dst) { *d = Dst(netip.MustParseAddr("192.0.2.10")) },
+	)(ctx)
+
+	want := User{name: "Bond"}
+	if gu := ValueOf[User](ctx); gu != want {
+		t.Errorf("different user %s", cmp.Diff(want, gu))
 	}
 }