Trying to reduce the boilerplate by using types as keys for context values.
An example of mutator (which can be used in `OnIntent`) is available as a test.

The performance impact of reflection are negligible:
```
goos: linux
goarch: amd64
pkg: github.com/TroutSoftware/rx
cpu: 12th Gen Intel(R) Core(TM) i7-12700
                 │ /tmp/before │             /tmp/after             │
                 │   sec/op    │   sec/op     vs base               │
ContextValues-20   1.433m ± 2%   1.518m ± 0%  +5.90% (p=0.000 n=10)
```